requirements.txt
beautifulsoup4
lxml
mecab-python3
spacy